Keep live example data inline
