+++ examples/live/src/lib.rs
@@ -1,14 +1,137 @@
 use pocopine::prelude::*;
 use serde::{Deserialize, Serialize};
 
-mod data;
-
-pub use data::query::{
-    create_post, list_posts, reset_posts, Post, PostDraft, POSTS_COLLECTION, POSTS_LIST_QUERY_TAG,
+#[cfg(pocopine_host)]
+use {
+    pocopine_events::{EventBackend, MemoryEventBackend},
+    std::sync::{
+        atomic::{AtomicU64, Ordering},
+        Mutex, OnceLock,
+    },
 };
 
+pub const POSTS_COLLECTION: &str = "posts";
+pub const POSTS_LIST_QUERY_TAG: &str = "posts:list";
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct Post {
+    pub id: String,
+    pub title: String,
+    pub body: String,
+    pub version: u64,
+}
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct PostDraft {
+    pub title: String,
+    pub body: String,
+}
+
 #[cfg(pocopine_host)]
-pub use data::query::{__create_post_route, __list_posts_route, __reset_posts_route, live_backend};
+static POSTS: OnceLock<Mutex<Vec<Post>>> = OnceLock::new();
+#[cfg(pocopine_host)]
+static NEXT_POST_ID: AtomicU64 = AtomicU64::new(3);
+#[cfg(pocopine_host)]
+static LIVE_BACKEND: OnceLock<MemoryEventBackend> = OnceLock::new();
+
+#[cfg(pocopine_host)]
+fn posts() -> &'static Mutex<Vec<Post>> {
+    POSTS.get_or_init(|| {
+        Mutex::new(vec![
+            Post {
+                id: "post_1".to_string(),
+                title: "Seeded post".to_string(),
+                body: "This row came from the process-local store.".to_string(),
+                version: 1,
+            },
+            Post {
+                id: "post_2".to_string(),
+                title: "Live updates".to_string(),
+                body: "Create a post in another tab and this list refreshes.".to_string(),
+                version: 1,
+            },
+        ])
+    })
+}
+
+#[cfg(pocopine_host)]
+pub fn live_backend() -> MemoryEventBackend {
+    LIVE_BACKEND.get_or_init(MemoryEventBackend::new).clone()
+}
+
+#[cfg(pocopine_host)]
+async fn publish_posts_invalidation(op: pocopine::live::LiveOp, key: impl Into<String>) {
+    let collection_draft = pocopine::live::LiveInvalidation::new(POSTS_COLLECTION, op)
+        .keys([key.into()])
+        .query_tags([POSTS_LIST_QUERY_TAG])
+        .into_draft();
+
+    publish_live_draft(collection_draft).await;
+
+    let query_draft = pocopine::live::query_tag_topic(POSTS_LIST_QUERY_TAG)
+        .and_then(|topic| pocopine::live::query_invalidated(topic, [POSTS_LIST_QUERY_TAG]));
+
+    publish_live_draft(query_draft).await;
+}
+
+#[cfg(pocopine_host)]
+async fn publish_live_draft(draft: pocopine_events::EventResult<pocopine_events::EventDraft>) {
+    let Ok(draft) = draft else {
+        return;
+    };
+
+    if let Err(err) = live_backend().publish(draft).await {
+        tracing::warn!(
+            target: "pocopine.log",
+            error = %err,
+            "failed to publish live posts invalidation"
+        );
+    }
+}
+
+#[pocopine::server(public)]
+pub async fn list_posts() -> ServerResult<Vec<Post>> {
+    let mut posts = posts()
+        .lock()
+        .map_err(|_| ServerError::App("posts store lock poisoned".to_string()))?
+        .clone();
+    posts.sort_by(|a, b| b.version.cmp(&a.version).then_with(|| b.id.cmp(&a.id)));
+    Ok(posts)
+}
+
+#[pocopine::server(public)]
+pub async fn create_post(draft: PostDraft) -> ServerResult<Post> {
+    let title = draft.title.trim();
+    let body = draft.body.trim();
+    if title.is_empty() || body.is_empty() {
+        return Err(ServerError::App("title and body are required".to_string()));
+    }
+
+    let next_id = NEXT_POST_ID.fetch_add(1, Ordering::Relaxed);
+    let post = Post {
+        id: format!("post_{next_id}"),
+        title: title.to_string(),
+        body: body.to_string(),
+        version: next_id,
+    };
+
+    posts()
+        .lock()
+        .map_err(|_| ServerError::App("posts store lock poisoned".to_string()))?
+        .push(post.clone());
+    publish_posts_invalidation(pocopine::live::LiveOp::Upsert, post.id.clone()).await;
+    Ok(post)
+}
+
+#[pocopine::server(public)]
+pub async fn reset_posts() -> ServerResult<()> {
+    posts()
+        .lock()
+        .map_err(|_| ServerError::App("posts store lock poisoned".to_string()))?
+        .clear();
+    publish_posts_invalidation(pocopine::live::LiveOp::Reset, POSTS_COLLECTION).await;
+    Ok(())
+}
 
 #[derive(Default, Serialize, Deserialize)]
 #[component(style = "live_board.css")]
